Move export to word function to ExamService
