+++ QuizManagement.WebApplication/QuizManagement.WebApplication.csproj
@@ -6,7 +6,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="DocumentFormat.OpenXml" Version="2.8.1" />
         <PackageReference Include="Microsoft.AspNetCore.App" />
         <PackageReference Include="Microsoft.AspNetCore.Connections.Abstractions" Version="2.1.3" />
         <PackageReference Include="Microsoft.AspNetCore.Session" Version="2.1.1" />
